// EscalafonFrontend/src/app/interfaces/instancia-gestion-educativa-descentralizada.ts
export interface InstanciaGestionEducativaDescentralizada {
    /**
     * @type { number } - iInstGeEduId
     */
    iInstGeEduId?: number;

    /**
     * @type { number } - iDirRegId
     */
    iDirRegId?: number;

    /**
     * @type { string } - cInstGeEduNombre
     */
    cInstGeEduNombre?: string;

}